refactor(components): extract task description and clean up task item variables
